+++ modules/@apostrophecms/color-field/ui/apos/logic/AposInputColor.js
@@ -16,6 +16,12 @@ export default {
     };
   },
   computed: {
+    isMicro() {
+      return this.modifiers.includes('micro');
+    },
+    isInline() {
+      return this.modifiers.includes('inline');
+    },
     pickerValue() {
       return this.next || this.defaultValue;
     },
@@ -32,13 +38,6 @@ export default {
     },
     options() {
       return this.field?.options || {};
-    },
-    valueLabel() {
-      if (this.next) {
-        return this.next;
-      } else {
-        return 'None Selected';
-      }
     },
     classList() {
       return [